WIP: splitting out impls into distinct traits
This commit starts to move the StructOpt trait into three smaller
traits (Parse, IntoApp, and FromArgMatches) as well as a convenience
triat which combines all three (Clap).

It also continues work on renaming structopt->clap

Finally, it starts to move all the impls into their own mods.
