+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "clap-derive"
+name = "clap_derive"
 description = "Custom Derive procedural macros for clap"
 version = "0.0.6"
 authors = [
@@ -11,20 +11,19 @@ license = "MIT/Apache-2.0"
 categories = ["command-line-interface", "development-tools::procedural-macro-helpers"]
 keywords = ["macro", "clap", "command-line", "structopt", "derive"]
 readme = "README.md"
-repository = "https://github.com/kbknapp/clap-derive"
+repository = "https://github.com/kbknapp/clap_derive"
 
 [lib]
 proc-macro = true
 
 [badges]
-travis-ci = { repository = "https://github.com/kbknapp/clap-derive" }
-appveyor = { repository = "https://github.com/kbknapp/clap-derive", service = "github" }
+travis-ci = { repository = "https://github.com/kbknapp/clap_derive" }
+appveyor = { repository = "https://github.com/kbknapp/clap_derive", service = "github" }
 
 [dependencies]
-clap = {features = ["unstable"], path = "../clap"}
+clap = {features = ["unstable"], path = "../clap-rs"}
 quote = "*"
 syn = "*"
-error-chain = "*"
 clippy    = { version = "~0.0.174", optional = true }
 proc-macro2 = "0.4"
