+++ rustfmt.toml
@@ -1,4 +1,2 @@
 format_strings = false
-chain_overflow_last = false
-same_line_if_else = true
 fn_single_line = true
