+++ src/derives/arg_enum.rs
@@ -1,19 +1,32 @@
+// Copyright 2018 Guillaume Pinot (@TeXitoi) <texitoi@texitoi.eu>,
+// Andrew Hobden (@hoverbear) <andrew@hoverbear.org>, and
+// Kevin Knapp (@kbknapp) <kbknapp@gmail.com>
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+use proc_macro2;
 use syn;
-use quote;
+// use syn::punctuated;
 
 use errors::*;
 
-pub(crate) fn impl_arg_enum(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
-    let from_str_block = impl_from_str(ast)?;
-    let variants_block = impl_variants(ast)?;
+pub fn impl_arg_enum(ast: &syn::DeriveInput) -> proc_macro2::TokenStream {
+    let from_str_block = impl_from_str(ast);
+    let variants_block = impl_variants(ast);
 
-    Ok(quote! {
+    quote! {
         #from_str_block
         #variants_block
-    })
+    }
 }
 
-fn impl_from_str(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
+fn impl_from_str(ast: &syn::DeriveInput) -> proc_macro2::TokenStream {
+    unimplemented!()
+    /*
     let ident = &ast.ident;
     let is_case_sensitive = ast.attrs.iter().any(|v| v.name() == "case_sensitive");
     let variants = variants(ast)?;
@@ -29,12 +42,10 @@ fn impl_from_str(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
 
     let for_error_message = strings.clone();
 
-    let condition_function_slice = [
-        match is_case_sensitive {
-            true => quote! { str::eq },
-            false => quote! { ::std::ascii::AsciiExt::eq_ignore_ascii_case },
-        },
-    ];
+    let condition_function_slice = [match is_case_sensitive {
+        true => quote! { str::eq },
+        false => quote! { ::std::ascii::AsciiExt::eq_ignore_ascii_case },
+    }];
     let condition_function = condition_function_slice.iter().cycle();
 
     Ok(quote! {
@@ -53,9 +64,12 @@ fn impl_from_str(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
             }
         }
     })
+    */
 }
 
-fn impl_variants(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
+fn impl_variants(ast: &syn::DeriveInput) -> proc_macro2::TokenStream {
+    unimplemented!()
+    /*
     let ident = &ast.ident;
     let variants = variants(ast)?
         .iter()
@@ -70,11 +84,14 @@ fn impl_variants(ast: &syn::DeriveInput) -> Result<quote::Tokens> {
             }
         }
     })
+    */
 }
 
-fn variants(ast: &syn::DeriveInput) -> Result<&Vec<syn::Variant>> {
-    match ast.body {
-        syn::Body::Enum(ref variants) => Ok(variants),
+/*
+fn variants(ast: &syn::DeriveInput) -> Result<&Vec<punctuated::Punctuated>> {
+    match ast.data {
+        syn::Data::Enum(ref data_enum) => Ok(data_enum.variants),
         _ => Err(ErrorKind::WrongBodyType("enum"))?,
     }
 }
+*/
