+++ src/errors.rs
@@ -1,4 +1,15 @@
-use proc_macro;
+// Copyright 2018 Guillaume Pinot (@TeXitoi) <texitoi@texitoi.eu>,
+// Andrew Hobden (@hoverbear) <andrew@hoverbear.org>, and
+// Kevin Knapp (@kbknapp) <kbknapp@gmail.com>
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+/*
+use proc_macro2;
 
 // Unfortunately `proc_macro` and `syn` don't have a good error handling story.
 error_chain! {
@@ -17,3 +28,5 @@ error_chain! {
         }
     }
 }
+
+*/
